+++ Sensorable/mobile/src/main/java/com/sensorable/MainActivity.java
@@ -1,7 +1,6 @@
 package com.sensorable;
 
 import android.app.Activity;
-import android.app.ActivityManager;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
@@ -20,16 +19,11 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
-import com.commons.DeviceType;
 import com.commons.LoginHelper;
 import com.commons.SensorTransmissionCoder;
 import com.commons.SensorableConstants;
+import com.commons.SensorableIntentFilters;
 import com.commons.SensorablePermissions;
-import com.commons.SensorsProvider;
-import com.commons.database.SensorMessageDao;
-import com.commons.database.SensorMessageEntity;
-import com.commons.devicesDetection.BluetoothDevicesProvider;
-import com.commons.devicesDetection.WifiDirectDevicesProvider;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
 import com.sensorable.activities.ActivitiesRegisterActivity;
@@ -37,19 +31,12 @@ import com.sensorable.activities.AdlSummaryActivity;
 import com.sensorable.activities.DetailedSensorsListActivity;
 import com.sensorable.activities.LocationOptionsActivity;
 import com.sensorable.activities.LoginActivity;
-import com.sensorable.services.AdlDetectionService;
-import com.sensorable.services.BackUpService;
 import com.sensorable.services.BluetoothDetectionService;
-import com.sensorable.services.EmpaticaTransmissionService;
-import com.sensorable.services.RegisterActivitiesService;
-import com.sensorable.services.SensorsProviderService;
-import com.sensorable.services.WearTransmissionService;
-import com.sensorable.utils.MobileDatabase;
-import com.sensorable.utils.MobileDatabaseBuilder;
+import com.sensorable.services.ManagerService;
 import com.sensorable.utils.MqttHelper;
 
 import java.util.ArrayList;
-import java.util.concurrent.ExecutorService;
+import java.util.NoSuchElementException;
 
 
 public class MainActivity extends AppCompatActivity {
@@ -60,12 +47,7 @@ public class MainActivity extends AppCompatActivity {
     private TextView userStateMessage;
     private TextView heartRateText, stepCounterText;
 
-    private SensorsProvider sensorsProvider;
-    private WifiDirectDevicesProvider wifiDirectProvider;
-
-    private SensorMessageDao sensorMessageDao;
-    private ExecutorService executor;
-    private BroadcastReceiver sensorDataReceiver;
+    private BroadcastReceiver displayDataReceiver;
 
 
     @Override
@@ -81,22 +63,10 @@ public class MainActivity extends AppCompatActivity {
 
         initializeAttributesFromUI();
 
-        initializeMobileDatabase();
         initializeSensorDataReceiver();
         initializeInfoReceiver();
 
-        initializeWearOsTranmissionService();
-        initializeEmpaticaTransmissionService();
-
-        initializeSensorsProviderService();
-        initializeAdlDetectionService();
-        initializeBackUpService();
-        initializeBluetoothDetectionService();
-
-        initializeRegisterActivitiesService();
-
-//        initializeSensors();
-//        initializeWifiDirectDetector();
+        initializeManagerService();
 
         MqttHelper.connect();
     }
@@ -164,44 +134,59 @@ public class MainActivity extends AppCompatActivity {
 
 
     private void initializeSensorDataReceiver() {
-        sensorDataReceiver =
+        displayDataReceiver =
                 new BroadcastReceiver() {
                     @Override
                     public void onReceive(Context context, Intent intent) {
                         Bundle b = intent.getBundleExtra(SensorableConstants.EXTRA_MESSAGE);
                         ArrayList<SensorTransmissionCoder.SensorMessage> arrayMessage = b.getParcelableArrayList(SensorableConstants.BROADCAST_MESSAGE);
-                        collectReceivedSensorData(arrayMessage);
 
-                        arrayMessage.forEach(sensorMessage -> {
-                            switch (sensorMessage.getDeviceType()) {
-                                case DeviceType.MOBILE:
-                                case DeviceType.WEAROS_RIGHT_HAND:
-                                case DeviceType.WEAROS_LEFT_HAND:
-                                case DeviceType.EMPATICA:
-                                    switch (sensorMessage.getSensorType()) {
-                                        case Sensor.TYPE_STEP_COUNTER:
-                                            int steps = Math.round(sensorMessage.getValue()[0]);
-                                            double stepsToKm = Math.round((steps * 0.65) / 10) / 100.0;
 
-                                            stepCounterText.setText(stepsToKm + " km");
-                                            useStateProgressBar.setProgress(steps);
-                                            userStateMessage.setText(getMessageBySteps(steps));
-                                            userStateSummary.setText(steps + "\n\n" + stepTarget);
-                                            break;
+                        // to display the average heart rate, we use streams
+                        try {
+                            double avgHeartRate = Math.floor(arrayMessage.stream()
+                                    .filter(sensor -> sensor.getSensorType() == Sensor.TYPE_HEART_RATE)
+                                    .mapToDouble(sensor -> sensor.getValue()[0])
+                                    .average()
+                                    .orElseThrow(NoSuchElementException::new));
 
-                                        case Sensor.TYPE_HEART_RATE:
-                                            int heartRate = Math.round(sensorMessage.getValue()[0]);
-                                            heartRateText.setText((heartRate == 0 ? "-" : heartRate) + " ppm");
-                                            break;
-                                    }
+                            heartRateText.setText((avgHeartRate == 0 ? "-" : avgHeartRate) + " ppm");
 
-                                    break;
-                            }
-                        });
+                            Log.i("DATA RECEIVER", "stream average heart rate" + avgHeartRate);
+
+                        } catch (NoSuchElementException e) {
+                            Log.e("DATA RECEIVER", "stream average has failed");
+                        }
+
+                        // to show the recorded steps we use the max operation of the stream method
+                        try {
+                            double recordedSteps = arrayMessage.stream()
+                                    .filter(s -> s.getSensorType() == Sensor.TYPE_STEP_COUNTER)
+                                    .mapToDouble(s -> s.getValue()[0])
+                                    .max()
+                                    .orElseThrow(NoSuchElementException::new);
+
+                            stepCounterText.setText(
+                                    Math.round((SensorableConstants.DISTANCE_OF_STEP_IN_M * recordedSteps) / 1000) + " Km"
+                            );
+
+                            userStateSummary.setText((int) recordedSteps + "\n" + stepTarget);
+
+                        } catch (NoSuchElementException e) {
+                            Log.e("DATA RECEIVER", "stream max steps has failed");
+
+                        }
                     }
-
-
                 };
+
+        LocalBroadcastManager.getInstance(this).
+                registerReceiver(displayDataReceiver, SensorableIntentFilters.EMPATICA_SENSORS);
+
+        LocalBroadcastManager.getInstance(this).
+                registerReceiver(displayDataReceiver, SensorableIntentFilters.WEAR_SENSORS);
+
+        LocalBroadcastManager.getInstance(this).
+                registerReceiver(displayDataReceiver, SensorableIntentFilters.MOBILE_SENSORS);
     }
 
     private String getMessageBySteps(int steps) {
@@ -217,29 +202,8 @@ public class MainActivity extends AppCompatActivity {
         }
 
         return msg;
-
     }
 
-
-    private void initializeMobileDatabase() {
-        MobileDatabase database = MobileDatabaseBuilder.getDatabase(this);
-        sensorMessageDao = database.sensorMessageDao();
-        executor = MobileDatabaseBuilder.getExecutor();
-    }
-
-    private void initializeWifiDirectDetector() {
-        wifiDirectProvider = new WifiDirectDevicesProvider(this);
-    }
-
-    private void initializeBluetoothDetectionService() {
-        if (!BluetoothDevicesProvider.isEnabledBluetooth()) {
-            BluetoothDevicesProvider.enableBluetooth(this);
-        } else {
-            if (!isMyServiceRunning(BluetoothDetectionService.class)) {
-                startService(new Intent(this, BluetoothDetectionService.class));
-            }
-        }
-    }
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
@@ -247,9 +211,7 @@ public class MainActivity extends AppCompatActivity {
             case SensorableConstants.REQUEST_ENABLE_BT:
                 Log.i("BLUETOOTH_PROVIDER", "on activity result for turn on bluetooth");
                 if (resultCode == Activity.RESULT_OK) {
-                    if (!isMyServiceRunning(BluetoothDetectionService.class)) {
-                        startService(new Intent(this, BluetoothDetectionService.class));
-                    }
+                    startService(new Intent(this, BluetoothDetectionService.class));
                 }
                 break;
             default:
@@ -260,89 +222,10 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    private void collectReceivedSensorData(ArrayList<SensorTransmissionCoder.SensorMessage> arrayMessage) {
-        // local database storing
-        executor.execute(() -> {
-            ArrayList<SensorMessageEntity> sensorMessageEntities = new ArrayList<>();
-            for (SensorTransmissionCoder.SensorMessage s : arrayMessage) {
-                sensorMessageEntities.add(s.toSensorDataMessage(LoginHelper.getUserCode(this)));
-            }
-            sensorMessageDao.insertAll(sensorMessageEntities);
-        });
-
-        sensorDataBuffer.addAll(arrayMessage);
-        sendSensorDataToAdlDetectionService();
+    private void initializeManagerService() {
+        startService(new Intent(this, ManagerService.class));
     }
 
-    private void collectReceivedSensorData(SensorTransmissionCoder.SensorMessage msg) {
-        executor.execute(() -> {
-            sensorMessageDao.insert(msg.toSensorDataMessage(LoginHelper.getUserCode(this)));
-        });
-
-        sensorDataBuffer.add(msg);
-        sendSensorDataToAdlDetectionService();
-    }
-
-    private void sendSensorDataToAdlDetectionService() {
-        if (sensorDataBuffer.size() >= SensorableConstants.COLLECTED_SENSOR_DATA_SIZE) {
-            Intent intent = new Intent(SensorableConstants.MOBILE_SENDS_SENSOR_DATA);
-            // You can also include some extra data.
-
-            Bundle adlBundle = new Bundle();
-            adlBundle.putParcelableArrayList(SensorableConstants.BROADCAST_MESSAGE, new ArrayList<>(sensorDataBuffer));
-
-            intent.putExtra(SensorableConstants.EXTRA_MESSAGE, adlBundle);
-            LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
-
-            // reset buffer
-            sensorDataBuffer.clear();
-        }
-    }
-
-    private boolean isMyServiceRunning(Class<?> serviceClass) {
-        ActivityManager manager = (ActivityManager) getSystemService(Context.ACTIVITY_SERVICE);
-        for (ActivityManager.RunningServiceInfo service : manager.getRunningServices(Integer.MAX_VALUE)) {
-            if (serviceClass.getName().equals(service.service.getClassName())) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private void initializeAdlDetectionService() {
-        if (!isMyServiceRunning(AdlDetectionService.class)) {
-            startService(new Intent(this, AdlDetectionService.class));
-        }
-    }
-
-    private void initializeBackUpService() {
-        if (!isMyServiceRunning(BackUpService.class)) {
-            startService(new Intent(this, BackUpService.class));
-        }
-    }
-
-    private void initializeEmpaticaTransmissionService() {
-        if (!isMyServiceRunning(EmpaticaTransmissionService.class)) {
-            startService(new Intent(this, EmpaticaTransmissionService.class));
-            // handle messages from our service to this activity
-        }
-
-        LocalBroadcastManager.getInstance(MainActivity.this).registerReceiver(
-                sensorDataReceiver,
-                new IntentFilter(SensorableConstants.EMPATICA_SENDS_SENSOR_DATA));
-    }
-
-    private void initializeWearOsTranmissionService() {
-        if (!isMyServiceRunning(WearTransmissionService.class)) {
-            // start new data transmission service to collect data from wear os
-            startService(new Intent(this, WearTransmissionService.class));
-        }
-
-        // handle messages from our service to this activity
-        LocalBroadcastManager.getInstance(MainActivity.this).registerReceiver(
-                sensorDataReceiver,
-                new IntentFilter(SensorableConstants.WEAR_SENDS_SENSOR_DATA));
-    }
 
     private void initializeAttributesFromUI() {
         userStateSummary = (Button) findViewById(R.id.userStateSummary);
@@ -365,28 +248,6 @@ public class MainActivity extends AppCompatActivity {
                 }, new IntentFilter(SensorableConstants.SERVICE_SENDS_INFO));
     }
 
-    private void initializeSensorsProviderService() {
-        if (!isMyServiceRunning(SensorsProviderService.class)) {
-            // start new data transmission service to collect data from wear os
-            startService(new Intent(this, SensorsProviderService.class));
-        }
-
-        // handle sensorMessages from sensors provider service and redirect this messages
-        // to the ADL detection service
-        // TODO: as you can see is redundant, we don't need to have this information if
-        //  the only task we are going to do is redirect it, we can send it directly
-        //  between services. So we need to create broadcast receivers in the sensors
-        //  provider service
-        LocalBroadcastManager.getInstance(MainActivity.this).registerReceiver(
-                sensorDataReceiver,
-                new IntentFilter(SensorableConstants.SENSORS_PROVIDER_SENDS_SENSORS));
-    }
-
-    private void initializeRegisterActivitiesService() {
-        if (!isMyServiceRunning(RegisterActivitiesService.class)) {
-            startService(new Intent(this, RegisterActivitiesService.class));
-        }
-    }
 
     private void initializeLogin() {
         if (!LoginHelper.isLogged(getApplicationContext())) {
